+++ doc/news.md
@@ -4,6 +4,7 @@ News, Updates, and Announcements
 - June 10, 2022: Version v0.10
 
       - USD 21.08 support, which is now the default in superbuild.
+      - Instance, Element, and Primitive ID buffers added
 
 - Feb 15, 2022: Version v0.9